renamed platforms -> chains
